Split tests and add calculatorpage test
